+++ rollup.config.js
@@ -2,7 +2,7 @@ import typescript from '@rollup/plugin-typescript';
 import pkg from './package.json';
 
 export default {
-  input: 'src/chart.ts',
+  input: 'src/index.ts',
   output: [{
     file: pkg.main,
     format: 'cjs',