+++ src/index.ts
@@ -1,0 +1,1 @@
+export { createChart } from './api/create-chart';