+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.0.2
+- Changed files structure
+- Added chart mouse move event for going back and forward in time
+- Added chart mouse wheel event for zoom
+
 ## v0.0.1
 - Simplified configuration (chartConfig function)
 - Added default themes